requirements.txt
streamlit
pandas
matplotlib
openpyxl
python-dateutil
joblib
sentence-transformers
scikit-learn